requirements.txt
autopep8==1.6.0
certifi==2022.12.7
charset-normalizer==2.0.7
click==8.0.3
et-xmlfile==1.1.0
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.24.2
openai==0.27.0
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.28.1
six==1.16.0
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
langchain==0.0.132
faiss-cpu==1.7.3
llama_index==0.5.12
PyPDF2==3.0.1
moviepy == 1.0.3
scipy == 1.10.1
soundfile == 0.12.1
moviepy
librosa
unstructured
